try to keep freeglut from calling exit, start making menu ops safer
also turn freeglut warnings/errors into cl WARN/ERROR
  Note that unwinding from errors might leave freeglut in a bad state,
  but still better than exiting the lisp. Might be able to recover by
  manually reloading the library, or leaving and restarting the main loop,
  but not guaranteed.

Glut specifies that most menu ops are not allowed while any menu is open.
This patch tries to handle the case where a menu is open when a window is
destroyed, since that is fairly easy to trigger. see #108 for general case.