add a runtime option to disable masking float traps
User code can use `%gl::with-float-traps-masked` to disable the per-call
masking within the dynamic scope of that macro. (not 100% final API, but
current plan is to export it without changes at some point.)

Building with masking completely disabled using existing :cl-opengl-no-masked-traps
feature is still a bit faster and slightly smaller, but disabling with the
runtime check is fairly close in performance.