Add option to defer calling glGetError, and optimize a bit for that use.
`%GL::WITH-DEFERRED-ERRORS (NAME)` can be used to defer calling `glGetError`
until the end of the dynamic scope of the body (assuming it it isn't deferred
by an enclosing scope) for cases where checking on every call is too slow.

To make debugging more convenient, it accepts an optional argument after NAME
to force checks on instead, so that can be set without removing the entire
form to get per-call checks on the next run.

API isn't final yet, but is expected to be exported without changes at some
point.