Rewrite defglextfun again
First step in #38 (optimize defglextfun) and #78 (support multiple contexts).
Also lets M-. continue working after extension functions are called.

Generator assigns a static index to each extension function.
A global vector stores functions that look up the function pointers and replace
the entry with a specialized function to call it.
The named functions are just thin inlined wrappers that call the correct entry
in the vector.